+++ src/experimental/erc7715/decorators/erc7715.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, test } from 'vitest'
 
 import { createClient } from '../../../clients/createClient.js'
 import { custom } from '../../../clients/transports/custom.js'
-import { walletActionsErc7115 } from './erc7115.js'
+import { walletActionsErc7715 } from './erc7715.js'
 
 const client = createClient({
   transport: custom({
@@ -20,10 +20,10 @@ const client = createClient({
       return null
     },
   }),
-}).extend(walletActionsErc7115())
+}).extend(walletActionsErc7715())
 
 test('default', async () => {
-  expect(walletActionsErc7115()(client)).toMatchInlineSnapshot(`
+  expect(walletActionsErc7715()(client)).toMatchInlineSnapshot(`
     {
       "issuePermissions": [Function],
     }